add frame_cap_description

+++ Database/database_manage.py
@@ -40,7 +40,8 @@ def get_ready():
     dir = ps.get_frame_cap_path()
     if os.listdir(dir): 
         for f in os.listdir(dir):
-            os.remove(os.path.join(dir, f))
+            if f != 'frame_cap_description.txt':
+                os.remove(os.path.join(dir, f))
     
     #criminal images
     dir = ps.get_assets_folder() + "criminal_images/"
